tester/requirements.txt
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
docker-py==0.5.0
requests==2.4.1
six==1.8.0
websocket-client==0.18.0
wsgiref==0.1.2
